fixing bugs

+++ src/controllers/Kebijakan.js
@@ -149,43 +149,12 @@ export const editKebijakan = async (req, res) => {
                 return res.status(403).json({msg: "Harus login dengan email dan username yang sesuai"});
             }       
         } else {
-            return res.status(404).json({msg: "Pemerintah tidak dapat mengupdate judul Kebijakan"})
+            return res.status(404).json({msg: "Rakyat tidak dapat mengupdate judul Kebijakan"})
         }
 
         res.status(200).json({msg: "Sukses mengedit judul Kebijakan"});
     } catch (error) {
         res.status(500).json({msg: error.message});
-    }
-}
-
-export const deleteKebijakanWithoutImage = async (req, res) => {
-    try {
-        const kebijakan = await Kebijakans.findOne({
-            where: {
-                kid: req.params.id
-            }
-        });
-
-        if (!kebijakan) {
-            return res.status(404).json({msg: "Kebijakan tidak ditemukan"});
-        }
-
-        if (req.role === "pemerintah") {
-            if (req.uid === kebijakan.userId) {
-                await Kebijakans.destroy({ 
-                    where:{
-                        // id: Kebijakan.id
-                        [Op.and]: [{kid: kebijakan.kid}, {userId: req.uid}]
-                    }
-                });
-            } else {
-                return res.status(403).json({msg: "Harus login dengan email dan username yang sesuai"});
-            }       
-        }
-
-        res.status(200).json({msg: "Sukses menghapus judul Kebijakan"});
-    } catch (error) {
-        res.status(500).json({msg: error.message}); 
     }
 }
 
@@ -203,8 +172,11 @@ export const deleteKebijakan = async (req, res) => {
 
         if (req.role === "pemerintah") {
             if (req.uid === kebijakan.userId) {
-                const filepath = `./public/images/kebijakan/${kebijakan.foto_data}`;
-                fs.unlinkSync(filepath);
+                const fotodata  = kebijakan.foto_data === "" && kebijakan.foto_url === "";
+                if (!fotodata){
+                    const filepath = `./public/images/kebijakan/${kebijakan.foto_data}`;
+                    fs.unlinkSync(filepath);
+                }
 
                 await Kebijakans.destroy({ 
                     where:{
@@ -223,56 +195,6 @@ export const deleteKebijakan = async (req, res) => {
     }
 }
 
-export const uploadImageKebijakanBaru = async (req, res) => {
-    const kebijakan = await Kebijakans.findOne({
-        where: {
-            kid: req.params.id
-        }
-    });
-
-    if (!kebijakan) {
-        return res.status(404).json({msg: "Tidak ditemukan kebijakan, harap buat kebijakan terlebih dahulu"});
-    }
-
-    let fileName = "";
-    if(req.files === null || req.files === undefined) {
-        fileName = kebijakan.foto_data;
-    } else {
-        const file = req.files.foto;
-        const fileSize = file.data.length;
-        const ext = path.extname(file.name);
-        fileName = file.md5 + ext;
-        const allowedType = ['.png','.jpg','.jpeg'];
-
-        if(!allowedType.includes(ext.toLowerCase())) {
-            return res.status(422).json({msg: "Foto yang anda masukkan tidak valid"});
-        }
-
-        if(fileSize > 50000000) {
-            return res.status(422).json({msg: "Ukuran foto harus kurang dari 50 MB"});
-        }
-
-        file.mv(`./public/images/kebijakan/${fileName}`, (err)=>{
-            if(err) {
-                return res.status(500).json({msg: err.message})
-            };
-        });
-    }
-
-    const url = `${req.protocol}://${req.get("host")}/images/kebijakan/${fileName}`;
-    
-    try {
-        await Kebijakans.update(
-            { foto_data: fileName, foto_url: url },
-            { where: { kid: req.params.id } }
-        );
-        res.status(200).json({msg: "Sukses memasukkan foto kebijakan"});
-    } catch (error) {
-        console.log(error.message);
-        res.status(400).json({msg: error.message});
-    }
-}
-
 export const editUploadImageKebijakan = async (req, res) => {
     const kebijakan = await Kebijakans.findOne({
         where: {
@@ -281,7 +203,7 @@ export const editUploadImageKebijakan = async (req, res) => {
     });
 
     if (!kebijakan) {
-        return res.status(404).json({msg: "Tidak ditemukan user, harap login terlebih dahulu"});
+        return res.status(404).json({msg: "Tidak ditemukan kebijakan, harap buat baru kebijakan terlebih dahulu"});
     }
 
     let fileName = "";
@@ -302,9 +224,11 @@ export const editUploadImageKebijakan = async (req, res) => {
             return res.status(422).json({msg: "Ukuran foto harus kurang dari 50 MB"});
         }
 
-        // const filepath = `./public/images/${product.image}`;
-        const filepath = `./public/images/kebijakan/${kebijakan.foto_data}`;
-        fs.unlinkSync(filepath);
+        const fotodata  = kebijakan.foto_data === null && kebijakan.foto_url === null;
+        if (!fotodata){
+            const filepath = `./public/images/kebijakan/${kebijakan.foto_data}`;
+            fs.unlinkSync(filepath);
+        }
 
         file.mv(`./public/images/kebijakan/${fileName}`, (err)=>{
             if(err) {
@@ -312,7 +236,7 @@ export const editUploadImageKebijakan = async (req, res) => {
             };
         });
     }
-    // const name = req.body.title;
+
     const url = `${req.protocol}://${req.get("host")}/images/kebijakan/${fileName}`;
     
     try {
