+++ src/controllers/Users.js
@@ -250,29 +250,6 @@ export const editPemerintah = async (req, res) => {
     }
 }
 
-export const deleteUserWithoutImage = async (req, res) => {
-    const user = await Users.findOne({
-        where: {
-            uid: req.params.id
-        }
-    });
-
-    if (!user) {
-        return res.status(404).json({msg: "User tidak ditemukan"});
-    }
-
-    try {
-        await Users.destroy({
-            where:{
-                uid: user.uid
-            }
-        });
-        res.status(200).json({msg: "Berhasil menghapus akun ini"});
-    } catch (error) {
-        res.status(400).json({msg: error.message});
-    }
-}
-
 export const deleteUser = async (req, res) => {
     const user = await Users.findOne({
         where: {
@@ -285,8 +262,11 @@ export const deleteUser = async (req, res) => {
     }
 
     try {
-        const filepath = `./public/images/users/${user.foto_data}`;
-        fs.unlinkSync(filepath);
+        const fotodata  = user.foto_data === "" && user.foto_url === "";
+        if (!fotodata){
+            const filepath = `./public/images/users/${user.foto_data}`;
+            fs.unlinkSync(filepath);
+        }
 
         await Users.destroy({
             where:{
@@ -295,55 +275,6 @@ export const deleteUser = async (req, res) => {
         });
         res.status(200).json({msg: "Berhasil menghapus akun ini"});
     } catch (error) {
-        res.status(400).json({msg: error.message});
-    }
-}
-
-export const uploadImageProfileBaru = async (req, res) => {
-    const users = await Users.findOne({
-        where: {
-            uid: req.params.id
-        }
-    });
-
-    if (!users) {
-        return res.status(404).json({msg: "Tidak ditemukan user, harap login terlebih dahulu"});
-    }
-
-    let fileName = "";
-    if(req.files === null || req.files === undefined) {
-        fileName = users.foto_data;
-    } else {
-        const file = req.files.foto;
-        const fileSize = file.data.length;
-        const ext = path.extname(file.name);
-        fileName = file.md5 + ext;
-        const allowedType = ['.png','.jpg','.jpeg'];
-
-        if(!allowedType.includes(ext.toLowerCase())) {
-            return res.status(422).json({msg: "Foto yang anda masukkan tidak valid"});
-        }
-
-        if(fileSize > 50000000) {
-            return res.status(422).json({msg: "Ukuran foto harus kurang dari 50 MB"});
-        }
-
-        file.mv(`./public/images/users/${fileName}`, (err)=>{
-            if(err) {
-                return res.status(500).json({msg: err.message})
-            };
-        });
-    }
-    const url = `${req.protocol}://${req.get("host")}/images/users/${fileName}`;
-    
-    try {
-        await Users.update(
-            {foto_data: fileName, foto_url: url },
-            {where: { uid: req.params.id } }
-        );
-        res.status(200).json({msg: "Sukses mengupdate foto profile"});
-    } catch (error) {
-        console.log(error.message);
         res.status(400).json({msg: error.message});
     }
 }
@@ -377,10 +308,12 @@ export const editUploadImageProfile = async (req, res) => {
             return res.status(422).json({msg: "Ukuran foto harus kurang dari 50 MB"});
         }
 
-        // const filepath = `./public/images/${product.image}`;
-        const filepath = `./public/images/users/${users.foto_data}`;
-        fs.unlinkSync(filepath);
-
+        const fotodata  = users.foto_data === null && users.foto_url === null;
+        if (!fotodata){
+            const filepath = `./public/images/users/${users.foto_data}`;
+            fs.unlinkSync(filepath);
+        }
+        
         file.mv(`./public/images/users/${fileName}`, (err)=>{
             if(err) {
                 return res.status(500).json({msg: err.message})
@@ -402,5 +335,3 @@ export const editUploadImageProfile = async (req, res) => {
     }
 
 }
-
-// export default { getUser, getUserById, createUser, editUser, deleteUser };
